+++ tg_bot/modules/animation.py
@@ -40,7 +40,7 @@ def fbi(bot: Bot, update: Update):
     
 __help__ = """
 - /police : Sends a police emoji animation. 
-- /fbi : Send O\_O animation
+- /fbi : Send O_O animation
 """
     
 POLICE_HANDLER = DisableAbleCommandHandler("police", police)
