+++ tg_bot/modules/sql/welcome_sql.py
@@ -77,6 +77,7 @@ DEFAULT_WELCOME_MESSAGES = [
     "Come for the Snail Racing, Stay for the Chimichangas!",
     "Who needs Google? You're everything we were searching for.",
     "This place must have free WiFi, cause I'm feeling a connection.",
+    "You shall not pass!",
     "Speak friend and enter.",
     "Welcome you are",
     "Welcome {first}, your princess is in another castle.",
@@ -102,7 +103,7 @@ DEFAULT_WELCOME_MESSAGES = [
     "Welcome {first}, if you’re a secret agent press 1, otherwise start a conversation",
     "{first}, I have a feeling we’re not in Kansas anymore.",
     "They may take our lives, but they’ll never take our {first}.",
-    "You had me at {first} joined the group.",
+    "You has me at {first} joined the group.",
     "Coast is clear! You can come out guys, it’s just {first}.",
     "Welcome {first}, Pay no attention to that guy lurking.",
     "Welcome {first}, May the force be with you.",
